+++ apps/ledger-live-desktop/src/live-common-set-supported-currencies.js
@@ -28,6 +28,7 @@ setSupportedCurrencies([
   "decred",
   "digibyte",
   "algorand",
+  "osmosis",
   "qtum",
   "bitcoin_gold",
   "komodo",
@@ -43,4 +44,5 @@ setSupportedCurrencies([
   "cosmos_testnet",
   "hedera",
   "cardano",
+  "filecoin",
 ]);