+++ apps/ledger-live-mobile/src/analytics/index.js
@@ -1,6 +1,0 @@
-// @flow
-import Track from "./Track";
-import TrackScreen from "./TrackScreen";
-
-export * from "./segment";
-export { Track, TrackScreen };