+++ apps/ledger-live-mobile/src/components/RootNavigator/BuyDeviceNavigator.tsx
@@ -5,6 +5,7 @@ import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 
 import { ScreenName } from "../../const";
 import { getStackNavigatorConfig } from "../../navigation/navigatorConfig";
+// eslint-disable-next-line import/no-cycle
 import GetDevice from "../../screens/GetDeviceScreen";
 import PurchaseDevice from "../../screens/PurchaseDevice";