+++ apps/ledger-live-mobile/src/screens/Manager/AppsList/AppUpdateStepper.tsx
@@ -1,12 +1,8 @@
 import React from "react";
 import { Trans } from "react-i18next";
-
 import { State } from "@ledgerhq/live-common/apps/index";
-
 import * as Animatable from "react-native-animatable";
-
 import { updateAllProgress } from "@ledgerhq/live-common/apps/logic";
-
 import { Flex, Text, ProgressBar } from "@ledgerhq/native-ui";
 
 const AnimatedFlex = Animatable.createAnimatableComponent(Flex);