+++ apps/ledger-live-mobile/src/screens/Manager/Device/index.tsx
@@ -44,10 +44,8 @@ type Props = {
   pendingInstalls: boolean;
   deviceInfo: DeviceInfo;
   device: Device;
-  setAppUninstallWithDependencies: (params: {
-    dependents: App[];
-    app: App;
-  }) => void;
+  setAppUninstallWithDependencies: (_: { dependents: App[]; app: App }) => void;
+  // eslint-disable-next-line no-unused-vars
   dispatch: (action: any) => void;
   appList: App[];
 };